backend/requirements.txt
fastapi
uvicorn
langchain-pinecone
langchain-openai
langchain-anthropic
pinecone-client
python-dotenv
pydantic